+++ lib/yt/collections/reports.rb
@@ -4,17 +4,18 @@ module Yt
   module Collections
     # @private
     class Reports < Base
-      DIMENSIONS = Hash.new({name: 'day', parse: ->(day) {Date.iso8601 day} }).tap do |hash|
-        hash[:traffic_source] = {name: 'insightTrafficSourceType', parse: ->(type) {TRAFFIC_SOURCES.key type} }
-        hash[:playback_location] = {name: 'insightPlaybackLocationType', parse: ->(type) {PLAYBACK_LOCATIONS.key type} }
-        hash[:embedded_player_location] = {name: 'insightPlaybackLocationDetail', parse: ->(url) {url} }
-        hash[:related_video] = {name: 'insightTrafficSourceDetail', parse: ->(video_id) { Yt::Video.new id: video_id, auth: @auth } }
-        hash[:video] = {name: 'video', parse: ->(video_id) { Yt::Video.new id: video_id, auth: @auth } }
-        hash[:playlist] = {name: 'playlist', parse: ->(playlist_id) { Yt::Playlist.new id: playlist_id, auth: @auth } }
-        hash[:device_type] = {name: 'deviceType', parse: ->(type) { type.downcase.to_sym } }
-        hash[:gender_age_group] = {name: 'gender,ageGroup', parse: ->(gender) { gender.downcase.to_sym }}
-        hash[:gender] = {name: 'gender', parse: ->(gender) { gender.downcase.to_sym } }
-        hash[:age_group] = {name: 'ageGroup', parse: ->(age_group) { age_group[3..-1] } }
+      DIMENSIONS = Hash.new({name: 'day', parse: ->(day, *values) {[Date.iso8601(day), values.last]} }).tap do |hash|
+        hash[:range] = {parse: ->(*values) { [:total, values.last]} }
+        hash[:traffic_source] = {name: 'insightTrafficSourceType', parse: ->(source, value) {[TRAFFIC_SOURCES.key(source), value]} }
+        hash[:playback_location] = {name: 'insightPlaybackLocationType', parse: ->(location, value) {[PLAYBACK_LOCATIONS.key(location), value]} }
+        hash[:embedded_player_location] = {name: 'insightPlaybackLocationDetail', parse: ->(url, value) {[url, value]} }
+        hash[:related_video] = {name: 'insightTrafficSourceDetail', parse: ->(video_id, value) { [Yt::Video.new(id: video_id, auth: @auth), value] } }
+        hash[:video] = {name: 'video', parse: ->(video_id, value) { [Yt::Video.new(id: video_id, auth: @auth), value] } }
+        hash[:playlist] = {name: 'playlist', parse: ->(playlist_id, value) { [Yt::Playlist.new(id: playlist_id, auth: @auth), value] } }
+        hash[:device_type] = {name: 'deviceType', parse: ->(type, value) { [type.downcase.to_sym, value] } }
+        hash[:gender_age_group] = {name: 'gender,ageGroup', parse: ->(gender, *values) { [gender.downcase.to_sym, *values] }}
+        hash[:gender] = {name: 'gender', parse: ->(gender, value) { [gender.downcase.to_sym, value] } }
+        hash[:age_group] = {name: 'ageGroup', parse: ->(age_group, value) { [age_group[3..-1], value] } }
       end
 
       # @see https://developers.google.com/youtube/analytics/v1/dimsmets/dims#Traffic_Source_Dimensions
@@ -47,7 +48,7 @@ module Yt
 
       attr_writer :metric
 
-      def within(days_range, dimension, try_again = true)
+      def within(days_range, dimension, type, try_again = true)
         @days_range = days_range
         @dimension = dimension
         if dimension == :gender_age_group # array of array
@@ -55,7 +56,7 @@ module Yt
             each{|gender, age_group, value| hash[gender][age_group[3..-1]] = value}
           end
         else
-          Hash[*flat_map{|value| [value.first, value.last]}]
+          Hash[*flat_map{|value| [value.first, type_cast(value.last, type)]}]
         end
       # NOTE: Once in a while, YouTube responds with 400 Error and the message
       # "Invalid query. Query did not conform to the expectations."; in this
@@ -64,13 +65,20 @@ module Yt
       # same query is a workaround that works and can hardly cause any damage.
       # Similarly, once in while YouTube responds with a random 503 error.
       rescue Yt::Error => e
-        try_again && rescue?(e) ? sleep(3) && within(days_range, dimension, false) : raise
+        try_again && rescue?(e) ? sleep(3) && within(days_range, dimension, type, false) : raise
       end
 
     private
 
+      def type_cast(value, type)
+        case [type]
+          when [Integer] then value.to_i if value
+          when [Float] then value.to_f if value
+        end
+      end
+
       def new_item(data)
-        [instance_exec(data.first, &DIMENSIONS[@dimension][:parse]), *data[1..-1]]
+        instance_exec *data, &DIMENSIONS[@dimension][:parse]
       end
 
       # @see https://developers.google.com/youtube/analytics/v1/content_owner_reports
@@ -87,7 +95,7 @@ module Yt
           params['start-date'] = @days_range.begin
           params['end-date'] = @days_range.end
           params['metrics'] = @metric.to_s.camelize(:lower)
-          params['dimensions'] = DIMENSIONS[@dimension][:name]
+          params['dimensions'] = DIMENSIONS[@dimension][:name] unless @dimension == :range
           params['max-results'] = 10 if @dimension == :video
           params['max-results'] = 200 if @dimension == :playlist
           params['max-results'] = 25 if @dimension == :embedded_player_location