+++ lib/yt/models/video.rb
@@ -373,65 +373,65 @@ module Yt
     ### ANALYTICS ###
 
       # @macro report_by_video_dimensions
-      has_report :views
+      has_report :views, Integer
 
       # @macro report_by_video_dimensions
-      has_report :estimated_minutes_watched
+      has_report :estimated_minutes_watched, Float
 
       # @macro report_by_gender_and_age_group
-      has_report :viewer_percentage
+      has_report :viewer_percentage, Float
 
       # @macro report_by_day
-      has_report :comments
+      has_report :comments, Integer
 
       # @macro report_by_day
-      has_report :likes
+      has_report :likes, Integer
 
       # @macro report_by_day
-      has_report :dislikes
+      has_report :dislikes, Integer
 
       # @macro report_by_day
-      has_report :shares
+      has_report :shares, Integer
 
       # @note This is not the total number of subscribers gained by the video’s
       #   channel, but the subscribers gained *from* the video’s page.
       # @macro report_by_day
-      has_report :subscribers_gained
+      has_report :subscribers_gained, Integer
 
       # @note This is not the total number of subscribers lost by the video’s
       #   channel, but the subscribers lost *from* the video’s page.
       # @macro report_by_day
-      has_report :subscribers_lost
+      has_report :subscribers_lost, Integer
 
       # @macro report_by_day
-      has_report :favorites_added
+      has_report :favorites_added, Integer
 
       # @macro report_by_day
-      has_report :favorites_removed
+      has_report :favorites_removed, Integer
 
       # @macro report_by_day
-      has_report :average_view_duration
+      has_report :average_view_duration, Float
 
       # @macro report_by_day
-      has_report :average_view_percentage
+      has_report :average_view_percentage, Float
 
       # @macro report_by_day
-      has_report :annotation_clicks
+      has_report :annotation_clicks, Integer
 
       # @macro report_by_day
-      has_report :annotation_click_through_rate
+      has_report :annotation_click_through_rate, Float
 
       # @macro report_by_day
-      has_report :annotation_close_rate
+      has_report :annotation_close_rate, Float
 
       # @macro report_by_day
-      has_report :earnings
+      has_report :earnings, Float
 
       # @macro report_by_day
-      has_report :impressions
+      has_report :impressions, Integer
 
       # @macro report_by_day
-      has_report :monetized_playbacks
+      has_report :monetized_playbacks, Integer
 
     ### STATISTICS ###