+++ spec/requests/as_content_owner/video_spec.rb
@@ -37,10 +37,36 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'earnings can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+        let(:keys) { range.values }
+
+        specify 'with the :by option set to :range' do
+          earnings = video.earnings range.merge by: :range
+          expect(earnings.size).to be 1
+          expect(earnings[:total]).to be_a Float
+        end
+      end
+
+      describe 'earnings can be grouped by day' do
+        let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
+        let(:keys) { range.values }
+
+        specify 'without a :by option (default)' do
+          earnings = video.earnings range
+          expect(earnings.keys).to eq range.values
+        end
+
+        specify 'with the :by option set to :day' do
+          earnings = video.earnings range.merge by: :day
+          expect(earnings.keys).to eq range.values
+        end
+      end
+
       describe 'views can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:views) { video.views_on 5.days.ago}
-          it { expect(views).to be_a Float }
+          it { expect(views).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -66,6 +92,16 @@ describe Yt::Video, :partner do
 
         specify 'with a given end (:to option)' do
           expect(video.views(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'views can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          views = video.views range.merge by: :range
+          expect(views.size).to be 1
+          expect(views[:total]).to be_an Integer
         end
       end
 
@@ -128,14 +164,14 @@ describe Yt::Video, :partner do
         specify 'with the :by option set to :device_type' do
           views = video.views range.merge by: :device_type
           expect(views.keys).to all(be_instance_of Symbol)
-          expect(views.values).to all(be_instance_of Float)
+          expect(views.values).to all(be_an Integer)
         end
       end
 
       describe 'comments can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:comments) { video.comments_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(comments).to be_a Float }
+          it { expect(comments).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -164,6 +200,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'comments can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          comments = video.comments range.merge by: :range
+          expect(comments.size).to be 1
+          expect(comments[:total]).to be_an Integer
+        end
+      end
+
       describe 'comments can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -182,7 +228,7 @@ describe Yt::Video, :partner do
       describe 'likes can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:likes) { video.likes_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(likes).to be_a Float }
+          it { expect(likes).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -211,6 +257,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'likes can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          likes = video.likes range.merge by: :range
+          expect(likes.size).to be 1
+          expect(likes[:total]).to be_an Integer
+        end
+      end
+
       describe 'likes can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -229,7 +285,7 @@ describe Yt::Video, :partner do
       describe 'dislikes can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:dislikes) { video.dislikes_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(dislikes).to be_a Float }
+          it { expect(dislikes).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -258,6 +314,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'dislikes can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          dislikes = video.dislikes range.merge by: :range
+          expect(dislikes.size).to be 1
+          expect(dislikes[:total]).to be_an Integer
+        end
+      end
+
       describe 'dislikes can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -276,7 +342,7 @@ describe Yt::Video, :partner do
       describe 'shares can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:shares) { video.shares_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(shares).to be_a Float }
+          it { expect(shares).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -305,6 +371,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'shares can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          shares = video.shares range.merge by: :range
+          expect(shares.size).to be 1
+          expect(shares[:total]).to be_an Integer
+        end
+      end
+
       describe 'shares can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -323,7 +399,7 @@ describe Yt::Video, :partner do
       describe 'gained subscribers can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:subscribers_gained) { video.subscribers_gained_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(subscribers_gained).to be_a Float }
+          it { expect(subscribers_gained).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -352,6 +428,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'gained subscribers can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          subscribers_gained = video.subscribers_gained range.merge by: :range
+          expect(subscribers_gained.size).to be 1
+          expect(subscribers_gained[:total]).to be_an Integer
+        end
+      end
+
       describe 'gained subscribers can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -370,7 +456,7 @@ describe Yt::Video, :partner do
       describe 'lost subscribers can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:subscribers_lost) { video.subscribers_lost_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(subscribers_lost).to be_a Float }
+          it { expect(subscribers_lost).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -399,6 +485,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'lost subscribers can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          subscribers_lost = video.subscribers_lost range.merge by: :range
+          expect(subscribers_lost.size).to be 1
+          expect(subscribers_lost[:total]).to be_an Integer
+        end
+      end
+
       describe 'lost subscribers can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -417,7 +513,7 @@ describe Yt::Video, :partner do
       describe 'added favorites can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:favorites_added) { video.favorites_added_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(favorites_added).to be_a Float }
+          it { expect(favorites_added).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -446,6 +542,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'added favorites can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          favorites_added = video.favorites_added range.merge by: :range
+          expect(favorites_added.size).to be 1
+          expect(favorites_added[:total]).to be_an Integer
+        end
+      end
+
       describe 'added favorites can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -464,7 +570,7 @@ describe Yt::Video, :partner do
       describe 'removed favorites can be retrieved for a specific day' do
         context 'in which the video was partnered' do
           let(:favorites_removed) { video.favorites_removed_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
-          it { expect(favorites_removed).to be_a Float }
+          it { expect(favorites_removed).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -490,6 +596,17 @@ describe Yt::Video, :partner do
 
         specify 'with a given end (:to option)' do
           expect(video.favorites_removed(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'removed favorites can be grouped by range' do
+        let(:id) { 'NeMlqbX2Ifg' }
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          favorites_removed = video.favorites_removed range.merge by: :range
+          expect(favorites_removed.size).to be 1
+          expect(favorites_removed[:total]).to be_an Integer
         end
       end
 
@@ -537,6 +654,16 @@ describe Yt::Video, :partner do
 
         specify 'with a given end (:to option)' do
           expect(video.estimated_minutes_watched(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'estimated minutes watched can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          minutes = video.estimated_minutes_watched range.merge by: :range
+          expect(minutes.size).to be 1
+          expect(minutes[:total]).to be_a Float
         end
       end
 
@@ -635,6 +762,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'average view duration can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          duration = video.average_view_duration range.merge by: :range
+          expect(duration.size).to be 1
+          expect(duration[:total]).to be_a Float
+        end
+      end
+
       describe 'average view duration can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -682,6 +819,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'average view percentage can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          percentage = video.average_view_percentage range.merge by: :range
+          expect(percentage.size).to be 1
+          expect(percentage[:total]).to be_a Float
+        end
+      end
+
       describe 'average view percentage can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -699,8 +846,8 @@ describe Yt::Video, :partner do
 
       describe 'impressions can be retrieved for a specific day' do
         context 'in which the video was partnered' do
-          let(:impressions) { video.impressions_on 20.days.ago}
-          it { expect(impressions).to be_a Float }
+          let(:impressions) { video.impressions_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
+          it { expect(impressions).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -729,6 +876,16 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'impressions can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          impressions = video.impressions range.merge by: :range
+          expect(impressions.size).to be 1
+          expect(impressions[:total]).to be_an Integer
+        end
+      end
+
       describe 'impressions can be grouped by day' do
         let(:range) { {since: 4.days.ago.to_date, until: 3.days.ago.to_date} }
         let(:keys) { range.values }
@@ -746,8 +903,8 @@ describe Yt::Video, :partner do
 
       describe 'monetized playbacks can be retrieved for a specific day' do
         context 'in which the video was partnered' do
-          let(:monetized_playbacks) { video.monetized_playbacks_on 20.days.ago}
-          it { expect(monetized_playbacks).to be_a Float }
+          let(:monetized_playbacks) { video.monetized_playbacks_on ENV['YT_TEST_PARTNER_VIDEO_DATE']}
+          it { expect(monetized_playbacks).to be_an Integer }
         end
 
         context 'in which the video was not partnered' do
@@ -773,6 +930,16 @@ describe Yt::Video, :partner do
 
         specify 'with a given end (:to option)' do
           expect(video.monetized_playbacks(to: date).keys.max).to eq date.to_date
+        end
+      end
+
+      describe 'monetized playbacks can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          monetized_playbacks = video.monetized_playbacks range.merge by: :range
+          expect(monetized_playbacks.size).to be 1
+          expect(monetized_playbacks[:total]).to be_an Integer
         end
       end
 
@@ -804,17 +971,27 @@ describe Yt::Video, :partner do
         end
       end
 
+      describe 'annotation clicks can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          annotation_clicks = video.annotation_clicks range.merge by: :range
+          expect(annotation_clicks.size).to be 1
+          expect(annotation_clicks[:total]).to be_an Integer
+        end
+      end
+
       describe 'annotation clicks can be grouped by day' do
         let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE'], until: Date.parse(ENV['YT_TEST_PARTNER_VIDEO_DATE']) + 5} }
 
         specify 'without a :by option (default)' do
           annotation_clicks = video.annotation_clicks range
-          expect(annotation_clicks.values).to all(be_instance_of Float)
+          expect(annotation_clicks.values).to all(be_an Integer)
         end
 
         specify 'with the :by option set to :day' do
           annotation_clicks = video.annotation_clicks range.merge by: :day
-          expect(annotation_clicks.values).to all(be_instance_of Float)
+          expect(annotation_clicks.values).to all(be_an Integer)
         end
       end
 
@@ -828,6 +1005,16 @@ describe Yt::Video, :partner do
 
         specify 'with a given start (:from option) and a given end (:to option)' do
           expect(video.annotation_click_through_rate(from: date, to: date_to).keys.min).to eq date.to_date
+        end
+      end
+
+      describe 'annotation click-through rate can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          rate = video.annotation_click_through_rate range.merge by: :range
+          expect(rate.size).to be 1
+          expect(rate[:total]).to be_a Float
         end
       end
 
@@ -855,6 +1042,16 @@ describe Yt::Video, :partner do
 
         specify 'with a given start (:from option) and a given end (:to option)' do
           expect(video.annotation_close_rate(from: date, to: date_to).keys.min).to eq date.to_date
+        end
+      end
+
+      describe 'annotation close rate can be grouped by range' do
+        let(:range) { {since: ENV['YT_TEST_PARTNER_VIDEO_DATE']} }
+
+        specify 'with the :by option set to :range' do
+          rate = video.annotation_close_rate range.merge by: :range
+          expect(rate.size).to be 1
+          expect(rate[:total]).to be_a Float
         end
       end